[Medium] [Array,Hash Table,Math,Counting,Number Theory] [248 ms] [20.5 MB]

+++ Number of Pairs of Interchangeable Rectangles/number-of-pairs-of-interchangeable-rectangles.go
@@ -1,5 +1,5 @@
 func interchangeableRectangles(rectangles [][]int) int64 {
-    ratio := map[float64]int{}
+    ratio := map[float64]int64{}
     for _, rectangle := range rectangles {
         ratio[float64(rectangle[0])/float64(rectangle[1])]++
     }
@@ -7,7 +7,7 @@ func interchangeableRectangles(rectangles [][]int) int64 {
     var res int64
     for _, v := range ratio {
         if v > 0 {
-            res += int64((v * (v-1)) / 2)
+            res += (v * (v-1)) / 2
         }
     }
